added ref: resource protocol
Allows specifying Spring Resources as literals referencing application properties. For example, ref:message will result in a resource that will read from application property message when an InputStream is requested (if message is not set, an IOException is thrown).

To use this feature, make sure that
com.github.dtreskunov.easyssl.ext.RefProtocolBeans
is getting imported by Spring. See included tests for an example.