release 2.2.2 with fix for a minor bug in 2.2.1
In 2.2.1, the definition of EasySslProperties bean was moved into a separate
class annotated with @Configuration. This broke code that was using an
explicit @Import annotation to configure source of bean definitions. This
commit adds an @Import to EasySslBeans so that the EasySslProperties bean
is included together with the beans in EasySslBeans.